+++ azure/terraform-module-storage-account/CHANGELOG.md
@@ -14,5 +14,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for deploying 0:n Storage Containers via the `containers` variable.
 - Support for RBAC role assignments against the Storage Account via the `role_assignments` variable.
 - Support for generating a Shared Access Signature (SAS) token via `sas_token_enabled` and `sas` variables.
-- Support for `network_rules` and `identity` optional blocks.
+- Support for `network_rules` and `identity` optional blocks. `network_rules` defaults to denying public network access except for trusted Azure services.
 - Outputs for `id`, `name`, `primary_blob_endpoint`, `primary_access_key`, `secondary_access_key`, `primary_connection_string`, `identity`, `containers`, `role_assignment_ids`, and `sas_token`.